+++ Övning5/Vehicles/MotorCycle.cs
@@ -2,7 +2,7 @@
 {
     internal class MotorCycle : Vehicle
     {
-        public MotorCycle(string id, string color, int wheels, float lenght) : base(id, color, wheels)
+        public MotorCycle(string id, string color, int wheels, int lenght) : base(id, color, wheels)
         {
             Id = id;
             Color = color;
@@ -13,6 +13,6 @@
         public string Id { get; }
         public string Color { get; }
         public int Wheels { get; }
-        public float Lenght { get; }
+        public int Lenght { get; }
     }
 }
